refactor(store): split git helpers
